+++ src/main/java/sonia/scm/auth/ldap/BindConnectionFailedException.java
@@ -23,10 +23,9 @@
  */
 package sonia.scm.auth.ldap;
 
-import org.apache.shiro.authc.AuthenticationException;
 
 @SuppressWarnings("squid:MaximumInheritanceDepth")
-public class BindConnectionFailedException extends AuthenticationException {
+public class BindConnectionFailedException extends LdapException {
 
   public BindConnectionFailedException(String message, Throwable cause) {
     super(message, cause);