+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Return empty list of groups on error instead of throwing exception ([#9](https://github.com/scm-manager/scm-ldap-plugin/pull/9))
+
 ## 2.0.0 - 2020-06-04
 ### Changed
 - Changeover to MIT license ([#3](https://github.com/scm-manager/scm-ldap-plugin/pull/3))